tokio_quiche

Function listen_with_capabilities

pub fn listen_with_capabilities<M>(
    sockets: impl IntoIterator<Item = QuicListener>,
    params: ConnectionParams<'_>,
    cid_generator: impl ConnectionIdGenerator<'static> + Clone,
    metrics: M,
) -> Result<Vec<QuicConnectionStream<M>>>
where
    M: Metrics,

Starts listening for inbound QUIC connections on the given QuicListeners.

Each socket starts a separate tokio task to process and route inbound packets. This task emits connections on the respective QuicConnectionStream after receiving the client’s QUIC initial and (optionally) validating its IP address.

The task shuts down when the returned stream is closed (or dropped) and all previously-yielded connections are closed.