tokio_quiche::http3

Module driver

An ApplicationOverQuic to build clients and servers on top of.

Modules

connection

Wrapper for running HTTP/3 connections.

Structs

H3Controller

Interface to communicate with a paired H3Driver.

H3Driver

A ready-made ApplicationOverQuic which can handle HTTP/3 and MASQUE. Depending on the DriverHooks in use, it powers either a client or a server.

IncomingH3Headers

HTTP/3 headers that were received on a stream.

NewClientRequest

An HTTP request sent using a ClientRequestSender to the ClientH3Driver.

RawPriorityValue

RequestSender

Sends H3Commands to an H3Driver. The sender is typed and internally wraps instances of T in the appropriate H3Command variant.

Enums

ClientH3Command

Commands accepted by ClientH3Driver.

ClientH3Event

Events produced by ClientH3Driver.

H3Command

H3Commands are sent by the H3Controller to alter the H3Driver’s state.

H3ConnectionError

The error type used internally in H3Driver.

H3Event

H3Events are produced by an H3Driver to describe HTTP/3 state updates.

InboundFrame

An InboundFrame is a data frame that was received from the peer over a quiche::h3::Connection. This is used by peers to send body or datagrams to the local task.

OutboundFrame

An OutboundFrame is a data frame that should be sent from a local task to a peer over a quiche::h3::Connection.

ServerH3Command

Commands accepted by ServerH3Driver.

ServerH3Event

Events produced by ServerH3Driver.

Type Aliases

ClientEventStream

Receives ClientH3Events from a ClientH3Driver. This is the control stream which describes what is happening on the connection, but does not transfer data.

ClientH3Controller

The H3Controller type paired with ClientH3Driver. See H3Controller for details.

ClientH3Driver

An H3Driver for a client-side HTTP/3 connection. See H3Driver for details. Emits ClientH3Events and expects ClientH3Commands for control.

ClientRequestSender

A RequestSender to send HTTP requests over a ClientH3Driver’s connection.

InboundFrameStream

Used by a local task to receive InboundFrames (data) on the stream or flow associated with this channel.

OutboundFrameSender

Used by a local task to send OutboundFrames to a peer on the stream or flow associated with this channel.

ServerEventStream

Receives ServerH3Events from a ServerH3Driver. This is the control stream which describes what is happening on the connection, but does not transfer data.

ServerH3Controller

The H3Controller type paired with ServerH3Driver. See H3Controller for details.

ServerH3Driver

An H3Driver for a server-side HTTP/3 connection. See H3Driver for details. Emits ServerH3Events and expects ServerH3Commands for control.