tokio_quiche::http3::driver

Struct H3Controller

pub struct H3Controller<H: DriverHooks> { /* private fields */ }

Interface to communicate with a paired H3Driver.

An H3Controller receives H3Events from its driver, which must be consumed by the application built on top of the driver to react to incoming events. The controller also allows the application to send ad-hoc H3Commands to the driver, which will be processed when the driver waits for incoming data.

Implementations

impl H3Controller<ClientHooks>

pub fn request_sender(&self) -> ClientRequestSender

Creates a NewClientRequest sender for the paired ClientH3Driver.

impl<H: DriverHooks> H3Controller<H>

pub fn event_receiver_mut(&mut self) -> &mut UnboundedReceiver<H::Event>

Gets a mut reference to the H3Event receiver for the paired H3Driver.

pub fn take_event_receiver(&mut self) -> UnboundedReceiver<H::Event>

Takes the H3Event receiver for the paired H3Driver.

pub fn cmd_sender(&self) -> RequestSender<H::Command, QuicCommand>

Creates a QuicCommand sender for the paired H3Driver.

pub fn send_goaway(&self)

Sends a GOAWAY frame to initiate a graceful connection shutdown.