tokio_quiche::http3::driver

Struct H3Driver

pub struct H3Driver<H: DriverHooks> { /* private fields */ }

A ready-made ApplicationOverQuic which can handle HTTP/3 and MASQUE. Depending on the DriverHooks in use, it powers either a client or a server.

Use the ClientH3Driver and ServerH3Driver aliases to access the respective driver types. The driver is passed into an I/O loop and communicates with the driver’s user (e.g., an HTTP client or a server) via its associated H3Controller. The controller allows the application to both listen for H3Events of note and send H3Commands into the I/O loop.

Implementations

impl<H: DriverHooks> H3Driver<H>

pub fn new(http3_settings: Http3Settings) -> (Self, H3Controller<H>)

Builds a new H3Driver and an associated H3Controller.

The driver should then be passed to InitialQuicConnection’s start method.