tokio_quiche::metrics

Module tokio_task

Instrumentation and metrics for spawned tokio tasks.

Currently, this is implemented by creating wrapper futures and wakers for the future inside of a spawned task. Ideally we would be able to move at least some of this work into tokio proper at some point, but this should be sufficient for now.

This does not rely on the tokio-metrics crate, as that has more overhead than we would like.

Functions

spawn

Spawn a potentially instrumented task.

spawn_with_killswitch

Spawn a potentially instrumented, long-lived task. Integrates with task-killswitch.