tokio_quiche::quic

Enum QuicCommand

pub enum QuicCommand {
    ConnectionClose(ConnectionShutdownBehaviour),
    Custom(Box<dyn FnOnce(&mut QuicheConnection) + Send + 'static>),
    Stats(Box<dyn FnOnce(SocketStats) + Send + 'static>),
}

A command to execute on a quiche::Connection in the context of an ApplicationOverQuic.

We expect most ApplicationOverQuic implementations (such as H3Driver) will provide some way to submit actions for them to take, for example via a channel. This enum may be accepted as part of those actions to inspect or alter the state of the underlying connection.

Variants

ConnectionClose(ConnectionShutdownBehaviour)

Close the connection with the given parameters.

Some packets may still be sent after this command has been executed, so the worker task may continue running for a bit. See quiche::Connection::close for details.

Custom(Box<dyn FnOnce(&mut QuicheConnection) + Send + 'static>)

Execute a custom callback on the connection.

Stats(Box<dyn FnOnce(SocketStats) + Send + 'static>)

Collect the current SocketStats from the connection.

Unlike QuicConnection::stats(), these statistics are not cached and instead are retrieved right before the command is executed.

Implementations

impl QuicCommand

pub fn execute(self, qconn: &mut QuicheConnection)

Consume the command and perform its operation on qconn.

This method should be called by ApplicationOverQuic implementations when they receive a QuicCommand to execute.