tokio_quiche::quic

Trait ConnectionHook

pub trait ConnectionHook {
    // Required method
    fn create_custom_ssl_context_builder(
        &self,
        settings: TlsCertificatePaths<'_>,
    ) -> Option<SslContextBuilder>;
}

A set of hooks executed at the level of a quiche::Connection.

Required Methods

fn create_custom_ssl_context_builder( &self, settings: TlsCertificatePaths<'_>, ) -> Option<SslContextBuilder>

Constructs an optional [SslContextBuilder].

This method allows full customization of quiche’s SSL context, for example to specify async callbacks during the QUIC handshake. It is called once per socket during initial setup, and then reused across all connections on that socket.

Only called if both the hook and TlsCertificatePaths are set in ConnectionParams.