tokio_quiche

Module socket

Network socket utilities and wrappers.

Structs

QuicListener

Wrapper around a [UdpSocket] for server-side QUIC connections.

Socket

A connected datagram socket with separate send and recv halves.

SocketCapabilities

Indicators of sockopts configured for a socket.

SocketCapabilitiesBuilder

Builder to enable Linux sockopts which improve QUIC performance.

Type Aliases

BoxedSocket

A type-erased variant of Socket with boxed Tx and Rx halves.