tokio_quiche::socket

Struct QuicListener

pub struct QuicListener {
    pub socket: UdpSocket,
    pub socket_cookie: u64,
    pub capabilities: SocketCapabilities,
}

Wrapper around a [UdpSocket] for server-side QUIC connections.

The wrapper carries socket-specific parameters, in contrast to the settings structs which apply to all sockets for a given QUIC server.

To create a QuicListener, you may either instantiate the struct yourself or use one of the TryFrom implementations.

Fields
socket: UdpSocket

The wrapped [tokio] socket.

socket_cookie: u64

An opaque value that is later passed to the ConnectionIdGenerator.

capabilities: SocketCapabilities

The SocketCapabilities to use for this socket.

By default, QuicListeners are constructed with all capabilities disabled. On Linux, you can use apply_max_capabilities() to (try to) enable all supported capabilities.

Implementations

impl QuicListener

pub fn apply_max_capabilities(&mut self)

Tries to enable all sockopts supported by the crate for this socket. See SocketCapabilities::apply_all_and_get_compatibility for details.