tokio_quiche::socket

Struct SocketCapabilitiesBuilder

pub struct SocketCapabilitiesBuilder<'s> { /* private fields */ }

Builder to enable Linux sockopts which improve QUIC performance.

Implementations

impl<'s> SocketCapabilitiesBuilder<'s>

pub fn new<S: AsFd>(socket: &'s S) -> Self

Creates a new sockopt builder for socket.

pub fn gso(&mut self) -> Result<()>

Enables UDP_SEGMENT, a generic segmentation offload (GSO).

GSO improves transmit performance by treating multiple sequential UDP packets as a single entity in the kernel. Segmentation into individual packets happens in the NIC, if it supports GSO. The parameter specifies the packet size.

pub fn check_udp_drop(&mut self) -> Result<()>

Enables SO_RXQ_OVFL, which reports dropped packets due to insufficient buffer space.

pub fn txtime(&mut self) -> Result<()>

Enables SO_TXTIME to control packet transmit timestamps for QUIC pacing.

pub fn rxtime(&mut self) -> Result<()>

Enables SO_TIMESTAMPNS, which records a wall-clock timestamp for each received packet.

pub fn gro(&mut self) -> Result<()>

Enables UDP_GRO, a generic receive offload (GRO).

GRO improves receive performance by allowing the kernel to yield multiple UDP packets in one recvmsg(2) call. It is the equivalent of GSO for the receive path.

pub fn ipv4_pktinfo(&mut self) -> Result<()>

Enables IP_PKTINFO to control the source IP in outbound IPv4 packets.

pub fn ipv4_recvorigdstaddr(&mut self) -> Result<()>

Enables IP_RECVORIGDSTADDR, which reports each packet’s real IPv4 destination address.

This can be different from the socket’s local address due to netfilter TPROXY rules or eBPF redirects.

pub fn ipv6_pktinfo(&mut self) -> Result<()>

Enables IPV6_RECVPKTINFO to control the source IP in outbound IPv6 packets.

pub fn ipv6_recvorigdstaddr(&mut self) -> Result<()>

Enables IPV6_RECVORIGDSTADDR, which reports each packet’s real IPv6 destination address.

This can be different from the socket’s local address due to netfilter TPROXY rules or eBPF redirects.

pub fn ip_mtu_discover_probe(&mut self) -> Result<()>

Sets IP_MTU_DISCOVER, to IP_PMTUDISC_PROBE, which disables kernel PMTUD and sets the DF (Don’t Fragment) flag.

pub fn ipv6_mtu_discover_probe(&mut self) -> Result<()>

Sets IPV6_MTU_DISCOVER, to IPV6_PMTUDISC_PROBE, which disables kernel PMTUD and sets the DF (Don’t Fragment) flag.

pub fn allows_nonlocal_source(&self) -> Result<bool>

Tests whether IP_FREEBIND or IP_TRANSPARENT are enabled for this socket.

Warning

These sockopts require elevated permissions to enable, so the builder will only check their status. If neither of them is enabled, the PKTINFO sockopts will cause errors when sending packets.

pub fn rcvmark(&mut self) -> Result<()>

pub fn finish(self) -> SocketCapabilities

Consumes the builder and returns the configured SocketCapabilities.