tokio_quiche::quic

Module raw

Helper to wrap existing quiche::Connections.

This is a low-level interface for users who either need to heavily customize the quiche::Connection beyond what is possible via the crate’s settings, or need more control over how to pass data into the connection.

Most use cases are much better served by our connect (for clients) or listen (for servers) API.

Structs

ConnCloseReceiver

Pollable receiver for connection closed notifications from a QUIC connection.

ConnWrapperResult

Result of manually wrapping a quiche::Connection in an InitialQuicConnection.

Functions

wrap_quiche_conn

Wraps an existing quiche::Connection in an InitialQuicConnection, bypassing the regular packet router workflow.