tokio_quiche::quic::raw

Function wrap_quiche_conn

pub fn wrap_quiche_conn<Tx, R, M>(
    quiche_conn: QuicheConnection,
    tx_socket: Socket<Arc<Tx>, R>,
    metrics: M,
) -> ConnWrapperResult<Tx, M>
where
    Tx: DatagramSocketSend + Send + 'static + ?Sized,
    M: Metrics,

Wraps an existing quiche::Connection in an InitialQuicConnection, bypassing the regular packet router workflow.

Connections wrapped in this way require the user to manually pass inbound packets via the channel returned in ConnWrapperResult. The passed tx_socket is only used to send outbound packets and to extract the endpoint’s addresses.

Note

This function does not attempt any I/O when wrapping the quiche::Connection. To start handshaking and consuming packets from the returned channel, use the methods on InitialQuicConnection.