tokio_quiche

Type Alias QuicConnectionStream

pub type QuicConnectionStream<M> = ReceiverStream<Result<InitialQuicConnection<UdpSocket, M>>>;

A stream of accepted InitialQuicConnections from a listen call.

Errors from processing the client’s QUIC initials can also be emitted on this stream. These do not indicate that the listener itself has failed.

Aliased Type

pub struct QuicConnectionStream<M> { /* private fields */ }